Fix musl/glibc cross-compile: cast libc::F_WRLCK to c_short
`libc::F_WRLCK` is `c_short` on macOS but `c_int` on glibc and musl,
while `flock.l_type` is always `c_short`. The earlier removal of the
cast was driven by macOS-only clippy advice; it broke the Linux release
build. The portable form is `as libc::c_short` — explicit about both
the target type and the reason it's needed.

+++ src/daemon/pidfile.rs
@@ -43,7 +43,9 @@ impl PidFile {
         // Try to acquire an exclusive non-blocking advisory lock.
         let fd = file.as_raw_fd();
         let mut flock: libc::flock = unsafe { std::mem::zeroed() };
-        flock.l_type = libc::F_WRLCK;
+        // libc::F_WRLCK is `c_short` on macOS, `c_int` on glibc/musl Linux;
+        // the cast is portable because `l_type` itself is always `c_short`.
+        flock.l_type = libc::F_WRLCK as libc::c_short;
         flock.l_whence = libc::SEEK_SET as i16;
         flock.l_start = 0;
         flock.l_len = 0;
